[DOP-8140] Split Hive.Dialect, .Slots and .*Options to separated submodules
