+++ docs/connection/db_connection/hive/connection.rst
@@ -1,0 +1,10 @@
+.. _hive-connection:
+
+Hive Connection
+===============
+
+.. currentmodule:: onetl.connection.db_connection.hive.connection
+
+.. autoclass:: Hive
+    :members: get_current, check
+    :member-order: bysource
