+++ docs/connection/db_connection/hive/execute.rst
@@ -1,0 +1,8 @@
+.. _hive-execute:
+
+Executing statements in Hive
+============================
+
+.. currentmodule:: onetl.connection.db_connection.hive.connection
+
+.. automethod:: Hive.execute
