+++ docs/connection/db_connection/hive/index.rst
@@ -1,0 +1,24 @@
+.. _hive:
+
+Hive
+====
+
+.. toctree::
+    :maxdepth: 1
+    :caption: Connection
+
+    connection
+
+.. toctree::
+    :maxdepth: 1
+    :caption: Operations
+
+    read
+    write
+    execute
+
+.. toctree::
+    :maxdepth: 1
+    :caption: For developers
+
+    slots
