+++ docs/connection/db_connection/hive/read.rst
@@ -1,0 +1,13 @@
+.. _hive-read:
+
+Reading from Hive
+=================
+
+There are 2 ways of distributed data reading from Hive:
+
+* Using :obj:`DBReader <onetl.db.db_reader.db_reader.DBReader>` with different :ref:`strategy`
+* Using :obj:`Hive.sql <onetl.connection.db_connection.hive.connection.Hive.sql>`
+
+.. currentmodule:: onetl.connection.db_connection.hive.connection
+
+.. automethod:: Hive.sql
