+++ docs/connection/db_connection/hive/slots.rst
@@ -1,0 +1,10 @@
+.. _hive-slots:
+
+Hive Slots
+==========
+
+.. currentmodule:: onetl.connection.db_connection.hive.slots
+
+.. autoclass:: HiveSlots
+    :members: normalize_cluster_name, get_known_clusters, get_current_cluster
+    :member-order: bysource
