+++ docs/connection/db_connection/hive/write.rst
@@ -1,0 +1,13 @@
+.. _hive-write:
+
+Writing to Hive
+===============
+
+For writing data to Hive, use :obj:`DBWriter <onetl.db.db_writer.db_writer.DBWriter>` with options below.
+
+.. currentmodule:: onetl.connection.db_connection.hive.options
+
+.. autopydantic_model:: HiveWriteOptions
+    :member-order: bysource
+    :model-show-field-summary: false
+    :field-show-constraints: false
